+++ R/compute_frequency_analysis.R
@@ -1,4 +1,4 @@
-# Copyright 2017 Province of British Columbia
+# Copyright 2018 Province of British Columbia
 # 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -212,11 +212,13 @@ compute_frequency_analysis <- function(data = NULL,
     ggplot2::geom_point()+
     ggplot2::xlab("Probability")+
     ggplot2::scale_x_continuous(trans = scales::probability_trans("norm", lower.tail = FALSE),
-                                breaks = prob_scale_points,
-                                sec.axis = ggplot2::sec_axis(trans = ~1/.,
-                                                             name = 'Return Period',
-                                                             breaks = c(1.01,1.1,2,5,10,20,100,1000),
-                                                             labels = function(x){ifelse(x < 2, x, round(x,0))}))+
+                                breaks = prob_scale_points#,
+                                # sec.axis = ggplot2::sec_axis(trans = ~1/.,
+                                #                              name = 'Return Period',
+                                #                              breaks = c(1.01,1.1,2,5,10,20,100,1000),
+                                #                              labels = function(x){ifelse(x < 2, x, round(x,0))}
+                                #                             )
+                                )+
     ggplot2::scale_color_brewer(palette = "Set1") +
     ggplot2::theme_bw() +
     ggplot2::labs(color = paste0('Events')) +